# gen/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Grand Kapital
  description: Взаимодействие с HTTP-сервисом, эмулирующий банковские транзакции и менеджмент пользовательский данных
...
paths:
  /users:
    get:
      tags:
        - users
      summary: Search users with filtering and pagination
      operationId: getUserPage
      description: |
        Search users with various filters. Available to all users.
        Filters:
        - dateOfBirth: returns users born after specified date
        - phone: exact match
        - name: starts with match (case insensitive)
        - email: exact match
      parameters:
        - name: pageNumber
          in: query
          required: true
          description: "page number"
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: pageSize
          in: query
          required: true
          description: "Number of users, returned to each page"
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: dateOfBirth
          in: query
          description: "Filter users born after this date (format: DD.MM.YYYY)"
          schema:
            type: string
            format: date
            example: "1990-05-15"
        - name: phone
          in: query
          description: "Filter by exact phone match (format: 79200234567)"
          schema:
            type: string
            pattern: '^7\d{10}$'
        - name: name
          in: query
          description: Filter by names starting with the value (case insensitive)
          schema:
            type: string
            minLength: 1
        - name: email
          in: query
          description: Filter by exact email match
          schema:
            type: string
      responses:
        200:
          description: List of users matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        400:
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}:
    get:
      tags:
        - users
      summary: Get user by ID
      operationId: getUserById
      description: Retrieve user information by ID
      parameters:
        - name: userId
          in: path
          description: user id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /transfer:
    post:
      security:
        - JWT: [ ]
      tags:
        - transfer
      summary: Transfer money from one to another user
      operationId: transfer
      description: ""
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        400:
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Page:
      description: "Page"
      type: object
      properties:
        number:
          description: "Page number"
          type: integer
          format: int32
        total:
          description: "Amount of pages"
          type: integer
          format: int32
        elementsTotal:
          description: "Total amount of elements on all pages"
          type: integer
          format: int64
        elements:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Иван Иванов
        dateOfBirth:
          type: string
          format: date
          example: "01.05.1993"
        emails:
          type: array
          items:
            type: string
          example: ["ivan.petrov@example.com"]
        phones:
          type: array
          items:
            type: string
            pattern: '^7\d{10}$'
          example: ["79200234567"]
        account:
            $ref: '#/components/schemas/Account'
    Account:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        balance:
          type: number
          format: decimal
          example: 1500.75
...
    Error:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          description: "message of error"
          example: "Invalid date format. Expected DD.MM.YYYY"
      required:
        - message
    Message:
      type: object
      properties:
        message:
          type: string
          description: "success message"
          example: "Success!"
      required:
        - message
...
    TransferRequest:
      type: object
      properties:
        toUserId:
          type: integer
          format: int64
          description: destination user
          example: 101
        amount:
          type: number
          format: double
          description: amount of money to transfer
          example: 99.9
      required:
        - toUserId
        - amount
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT